Update Makefile to filter out some includes for doc's generation.  Specifically I am using it to remove 3rd party code from the docs.
Removed the c++format library from docs per above.

Updated TreeAnalyzer to allow multiple event loops. The class doc should now show the structure of a TreeAnalyzer job.  Multiple event loops can be used.  I am not sure how I feel about this. It is experimental and may be removed in favor of a different solution in the future. For instance a better way of running multiple Task's in the same program call.

Updated reporter to have an explicit close function. Useful in root scripts because the class destructor doesnt seem to be called.

Signed-off-by: jdbrice <jdb12@rice.edu>